feat: trial next interface
