--- examples/playwright-orca/tests/firefox/firefox.itemText.snapshot.json
@@ -1,10 +1,0 @@
-[
-  "Guidepup heading level 1",
-  "Screen Reader A11y Workflows",
-  "Full Control heading level 3",
-  "Mirrors Real User Experience heading level 3",
-  "Framework Agnostic heading level 3",
-  "Run with Jest, with Playwright, as an independent script, no vendor lock-in.",
-  "Get Started link",
-  "GitHub link"
-]
