--- examples/playwright-orca/tests/firefox/firefox.spokenPhrase.snapshot.json
@@ -1,10 +1,0 @@
-[
-  "heading level 1 Guidepup",
-  "Screen Reader A11y Workflows",
-  "heading level 3 Full Control",
-  "heading level 3 Mirrors Real User Experience",
-  "heading level 3 Framework Agnostic",
-  "Run with Jest, with Playwright, as an independent script, no vendor lock-in.",
-  "link Get Started",
-  "link GitHub"
-]
