--- examples/playwright-orca/tests/headerNavigation.ts
@@ -1,66 +1,0 @@
-import { delay } from "../../../src/delay";
-import { log } from "../../log";
-import { OrcaPlaywright } from "../orca-test";
-import { Page } from "@playwright/test";
-
-const MAX_NAVIGATION_LOOP = 10;
-
-export async function headerNavigation({
-  page,
-  orca,
-}: {
-  page: Page;
-  orca: OrcaPlaywright;
-}) {
-  // Navigate to Guidepup Website 🎉
-  log("Navigating to URL: https://www.guidepup.dev.");
-  await page.goto("https://www.guidepup.dev", {
-    waitUntil: "load",
-  });
-
-  // Wait for page to be ready and interact 🙌
-  const header = page.locator("h1");
-  await header.waitFor();
-  await delay(500);
-
-  // Make sure interacting with the web content
-  await orca.navigateToWebContent();
-  await delay(500);
-
-  let headingCount = 0;
-
-  // Move across the headings using Orca 🔎
-  while (
-    !(await orca.itemText()).includes("Framework Agnostic") &&
-    headingCount <= MAX_NAVIGATION_LOOP
-  ) {
-    headingCount++;
-
-    log(`Performing command: "H" - "Find the next heading"`);
-    await orca.nextHeading();
-    log(`Screen reader output: "${await orca.lastSpokenPhrase()}".`);
-  }
-
-  let tabCount = 0;
-
-  // Move across text and buttons using Orca
-  while (
-    !(await orca.itemText()).includes("GitHub") &&
-    tabCount <= MAX_NAVIGATION_LOOP
-  ) {
-    tabCount++;
-
-    log(`Performing command: "Down Arrow"`);
-    await orca.next();
-    log(`Screen reader output: "${await orca.lastSpokenPhrase()}".`);
-  }
-
-  log(`Performing command: "Up Arrow"`);
-  await orca.previous();
-  log(`Screen reader output: "${await orca.lastSpokenPhrase()}".`);
-
-  // Navigate to the Guidepup Getting Started docs
-  log(`Performing command: "Enter"`);
-  await orca.act();
-  log(`Screen reader output: "${await orca.lastSpokenPhrase()}".`);
-}
